+++ examples/contracts/public_currency/test/PublicCurrencyQuorumTest.sol
@@ -7,7 +7,7 @@ import "dao-currency/test/MockCurrencyDatabase.sol";
 contract PublicCurrencyQuorumTest is PublicCurrencyBase {
 
     function testCreateQuorumBallotSuccess() {
-        var mdb = new MockUserDatabase(block.timestamp, true, TEST_NUM_ELIGIBLE_VOTERS);
+        var mdb = new MockUserDatabase(block.timestamp, true, true, TEST_NUM_ELIGIBLE_VOTERS);
         var pc = new PublicCurrency(TEST_ADDRESS, mdb);
         pc.createQuorumBallot(1).assertNoError("createQuorumBallot returned error");
 
@@ -29,7 +29,7 @@ contract PublicCurrencyQuorumTest is PublicCurrencyBase {
     }
 
     function testCreateQuorumBallotFailQuorumIsTheSame() {
-        var mdb = new MockUserDatabase(block.timestamp, true, TEST_NUM_ELIGIBLE_VOTERS);
+        var mdb = new MockUserDatabase(block.timestamp, true, true, TEST_NUM_ELIGIBLE_VOTERS);
         var pc = new PublicCurrency(TEST_ADDRESS, mdb);
         pc.createQuorumBallot(TEST_QUORUM).assertErrorsEqual(INVALID_PARAM_VALUE, "createQuorumBallot returned the wrong error");
 
@@ -38,7 +38,7 @@ contract PublicCurrencyQuorumTest is PublicCurrencyBase {
     }
 
     function testCreateQuorumBallotFailNotUser() {
-        var mdb = new MockUserDatabase(0, false, 0);
+        var mdb = new MockUserDatabase(0, false, true, 0);
         var pc = new PublicCurrency(TEST_ADDRESS, mdb);
         pc.createQuorumBallot(1).assertErrorsEqual(RESOURCE_NOT_FOUND, "createQuorumBallot returned the wrong error");
 
@@ -47,13 +47,13 @@ contract PublicCurrencyQuorumTest is PublicCurrencyBase {
     }
 
     function testSetQuorumFailCallerNotBallot() {
-        var mdb = new MockUserDatabase(0, false, 0);
+        var mdb = new MockUserDatabase(0, false, true, 0);
         var pc = new PublicCurrency(TEST_ADDRESS, TEST_ADDRESS_2);
         pc.setQuorum(TEST_QUORUM).assertErrorsEqual(ACCESS_DENIED, "setQuorum returned no 'access denied' error");
     }
 
     function testQuorumBallotSuccessVotePassed() {
-        var mdb = new MockUserDatabase(block.timestamp, true, TEST_NUM_ELIGIBLE_VOTERS);
+        var mdb = new MockUserDatabase(block.timestamp, true, true, TEST_NUM_ELIGIBLE_VOTERS);
         var mcd = new MockCurrencyDatabase();
         var pc = new PublicCurrency(mcd, mdb);
         pc.createQuorumBallot(1).assertNoError("createQuorumBallot returned error");
@@ -71,7 +71,7 @@ contract PublicCurrencyQuorumTest is PublicCurrencyBase {
     }
 
     function testQuorumBallotSuccessVoteDidNotPass() {
-        var mdb = new MockUserDatabase(block.timestamp, true, TEST_NUM_ELIGIBLE_VOTERS);
+        var mdb = new MockUserDatabase(block.timestamp, true, true, TEST_NUM_ELIGIBLE_VOTERS);
         var mcd = new MockCurrencyDatabase();
         var pc = new PublicCurrency(mcd, mdb);
         pc.createQuorumBallot(1).assertNoError("createQuorumBallot returned error");